Exception handling - delete

+++ src/main/java/com/lucas/project/services/UserService.java
@@ -2,8 +2,11 @@ package com.lucas.project.services;
 
 import com.lucas.project.entities.User;
 import com.lucas.project.repositories.UserRepository;
+import com.lucas.project.services.exceptions.DatabaseException;
 import com.lucas.project.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -29,7 +32,20 @@ public class UserService {
     }
 
     public void delete(Long id){
-        userRepository.deleteById(id);
+        try{
+           if(userRepository.existsById(id)){
+               userRepository.deleteById(id);
+           }
+           else {
+               throw new ResourceNotFoundException(id);
+           }
+        }
+        catch (EmptyResultDataAccessException e){
+            throw new ResourceNotFoundException(id);
+        }
+        catch (DataIntegrityViolationException e){
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User user) {
